Initial commit
The supplied text document and spreadsheet have been merged. This will need to be cleaned up and missing data filled in.

The merged CSV file is pulled through and displayed (only showing the name, url, type, focus and permissions currently).

Basic free text search has been added which filters on name and URL.

The dropdown filters need linking up to the data to filter correctly, however a data structure should be agreed on first before this is done.